move client-side components to dedicated entry points
