+++ packages/react/src/dynamic/index.ts
@@ -1,0 +1,6 @@
+'use client'
+
+export { TokenIcon } from './TokenIcon'
+export { NetworkIcon } from './NetworkIcon'
+export { WalletIcon } from './WalletIcon'
+export { ExchangeIcon } from './ExchangeIcon'
