+++ packages/react/src/index.ts
@@ -1,10 +1,8 @@
 export * from './icons'
-export * from './TokenIcon'
-export * from './NetworkIcon'
-export * from './WalletIcon'
-export * from './ExchangeIcon'
 export * from './types'
+export * from './BaseIcon'
 export * as tokenIcons from './icons/tokens'
 export * as networkIcons from './icons/networks'
 export * as walletIcons from './icons/wallets'
 export * as exchangeIcons from './icons/exchanges'
+export * from './dynamic'
