+++ packages/react/tsup.config.ts
@@ -3,6 +3,7 @@ import { defineConfig } from 'tsup'
 export default defineConfig({
   entry: {
     index: 'src/index.ts',
+    dynamic: 'src/dynamic/index.ts',
     'icons/tokens': 'src/icons/tokens/index.ts',
     'icons/networks': 'src/icons/networks/index.ts',
     'icons/wallets': 'src/icons/wallets/index.ts',
